pyproject.toml
[project]
name = "github_graphql_mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "mcp[cli]>=1.6.0",
    "graphql-core>=3.2.6",
    "requests",
    "python-dotenv"
]

[tool.poetry]
package-mode = false

[tool.setuptools]
package-dir = {"" = "src"}